refactor: Cleanup constants
