+++ tools/ui/src/lib/constants/control-actions.constants.ts
@@ -3,5 +3,3 @@
 export const CONTROL_ACTION = {
 	END_REASONING: 'reasoning_end'
 } as const;
-
-export type ControlAction = (typeof CONTROL_ACTION)[keyof typeof CONTROL_ACTION];
