--- tools/ui/src/lib/constants/floating-ui-constraints.constants.ts
@@ -1,2 +1,0 @@
-export const VIEWPORT_GUTTER = 8;
-export const MENU_OFFSET = 6;
