+++ tools/ui/src/lib/constants/index.ts
@@ -18,14 +18,13 @@ export * from './content-detection.constants';
 export * from './built-in-tools.constants';
 export * from './cache.constants';
 export * from './chat-form.constants';
-export * from './chat-commands.constants';
 export * from './cli-flags.constants';
 export * from './code-block.constants';
 export * from './context-keys.constants';
 export * from './control-actions.constants';
 export * from './css-classes.constants';
-export * from './floating-ui-constraints.constants';
 export * from './formatters.constants';
+export * from './headers.constants';
 export * from './key-value-pairs.constants';
 export * from './latex-protection.constants';
 export * from './literal-html.constants';
@@ -45,7 +44,6 @@ export * from './path-display.constants';
 export * from './model-id.constants';
 export * from './model-loading.constants';
 export * from './precision.constants';
-export * from './processing-info.constants';
 export * from './pwa.constants';
 export * from './routes.constants';
 export * from './sandbox.constants';
@@ -56,11 +54,8 @@ export * from './stream.constants';
 export * from './supported-file-types.constants';
 export * from './table-html-restorer.constants';
 export * from './title-generation.constants';
-export * from './tools.constants';
-export * from './tooltip-config.constants';
 export * from './ui.constants';
 export * from './uri-template.constants';
 export * from './url.constants';
-export * from './viewport.constants';
 export * from './working-directory.constants';
 export * from './read-media';
