--- tools/ui/src/lib/constants/processing-info.constants.ts
@@ -1,8 +1,0 @@
-export const PROCESSING_INFO_TIMEOUT = 2000;
-
-/**
- * Statistics units labels
- */
-export const STATS_UNITS = {
-	TOKENS_PER_SECOND: 't/s'
-} as const;
