+++ tools/ui/src/lib/constants/stream.constants.ts
@@ -2,6 +2,10 @@
 // while the tab was hidden. covers brief background pauses without thrashing live streams
 export const STREAM_VISIBILITY_KICK_MS = 3000;
 
+// separator joining a conversation id and its per-model stream identity
+// suffix (conv::model) used by the server side replay buffer
+export const CONVERSATION_ID_SEPARATOR = '::';
+
 /**
  * Server-sent events wire format, shared by the chat stream and the
  * /models/sse status feed (text/event-stream).
