--- tools/ui/src/lib/constants/tooltip-config.constants.ts
@@ -1,1 +1,0 @@
-export const TOOLTIP_DELAY_DURATION = 500;
