+++ tools/ui/src/lib/hooks/use-chat-form-pickers.svelte.ts
@@ -1,6 +1,7 @@
-import { getChatCommands, PROMPT_TRIGGER_PREFIX } from '$lib/constants';
+import { PROMPT_TRIGGER_PREFIX } from '$lib/constants';
 import { ChatFormCommandAction, KeyboardKey } from '$lib/enums';
 import type { ChatFormCommand } from '$lib/types';
+import { getChatCommands } from '$lib/utils';
 import {
 	type CommandDismissSnapshot,
 	findCommandToken,
