+++ tools/ui/src/lib/types/index.ts
@@ -56,7 +56,9 @@ export type {
 	AttachmentDisplayItemsOptions,
 	FileProcessingResult,
 	FileMentionEntry,
-	ChatFormCommand
+	ChatFormCommand,
+	ChatCommandsOptions,
+	ControlAction
 } from './chat.d';
 
 // Database types
