+++ tools/ui/src/lib/utils/stream-identity.ts
@@ -1,3 +1,5 @@
+import { CONVERSATION_ID_SEPARATOR } from '$lib/constants';
+
 /**
  * Build the conversation identity used by the server side replay buffer.
  *
@@ -11,5 +13,5 @@ export function streamIdentity(conversationId: string, model?: string | null): s
 
 	if (!model) return conversationId;
 
-	return `${conversationId}::${model}`;
+	return `${conversationId}${CONVERSATION_ID_SEPARATOR}${model}`;
 }
